# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phishing Simulation Platform API",
        "description": "Comprehensive API for managing phishing simulation campaigns, employee training, and security awareness programs.",
        "contact": {
            "email": "support@phishingsim.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/payment/checkout": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Initialize payment checkout",
                "description": "Initialize a payment checkout process for a subscription plan",
                "operationId": "2414a7c0f1c0a430af10755ef7876abf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "description": "ID of the subscription plan",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment initialized successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment initialized successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "txn_123456789abcdef"
                                                },
                                                "checkout_url": {
                                                    "type": "string",
                                                    "example": "https://payment-gateway.com/checkout/txn_123456789abcdef"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                },
                                                "plan": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Basic"
                                                        },
                                                        "price": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 10
                                                        },
                                                        "employee_limit": {
                                                            "type": "integer",
                                                            "example": 50
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "Validation error message"
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}